ml-backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
httpx>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0
python-multipart>=0.0.6
langchain-community>=0.0.8
langchain-core>=0.1.5
langchain-huggingface>=0.0.3
google-generativeai>=0.3.0
Pillow>=10.0.0
